+++ js/data.js
@@ -1,0 +1,52 @@
+const getSettings = () => ({
+  DESCRIPTION_MAX_COUNT: 25,
+  LIKES_MIN_COUNT: 15,
+  LIKES_MAX_COUNT: 200,
+  COMMENTS_MIN_COUNT: 0,
+  COMMENTS_MAX_COUNT: 30,
+  COMMENT_AUTHORS_MAX_COUNT: 6
+});
+
+const getCommentMessages = () => ({
+  good: ['Всё отлично!'],
+  neutral: ['В целом всё неплохо. Но не всё.'],
+  bad: [
+    'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально',
+    'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+    'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+    'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+  ]
+});
+
+const getPhotoDescriptions = () => ([
+  'Момент вечности',
+  'Ловлю мгновение',
+  'Красота в каждом кадре',
+  'Магия момента',
+  'Мир, созданный изображением',
+  'История без слов',
+  'Эмоции в фокусе',
+  'Одно мгновение, тысячи впечатлений',
+  'Когда фото говорит больше слов',
+  'Вдохновение сквозь объектив',
+  'Мир глазами фотографа',
+  'Фотография – это язык сердца',
+  'Моменты, которые запоминаются',
+  'Фотография, чтобы запомнить'
+]);
+
+const getCommentAuthors = () => ({
+  '1': { name: 'Виктор' },
+  '2': { name: 'Кирилл' },
+  '3': { name: 'Игорь' },
+  '4': { name: 'Мария' },
+  '5': { name: 'Татьяна' },
+  '6': { name: 'Наталья' },
+});
+
+export {
+  getSettings,
+  getCommentMessages,
+  getPhotoDescriptions,
+  getCommentAuthors
+};
